requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0 